rename location for bulkEnrichment
