+++ src/endpoints/index.ts
@@ -1,5 +1,5 @@
 import autocomplete from './autocomplete';
-import bulk from './bulk';
+import bulkEnrichment from './bulkEnrichment';
 import bulkRetrieve from './bulkRetrieve';
 import cleaner from './cleaner';
 import enrichment from './enrichment';
@@ -9,4 +9,4 @@ import retrieve from './retrieve';
 import search from './search';
 import skill from './skill';
 
-export { autocomplete, bulk, bulkRetrieve, cleaner, enrichment, identify, jobTitle, retrieve, search, skill };
+export { autocomplete, bulkEnrichment, bulkRetrieve, cleaner, enrichment, identify, jobTitle, retrieve, search, skill };
